requirements.txt
flask>=3.0.0
requests>=2.31.0
openpyxl>=3.1.2
pdfplumber>=0.10.0
pywebview>=5.0.0
pyinstaller>=6.0.0
keyring>=25.0.0
Pillow>=10.0.0
